Response/View Counts
Added a new feature to track with each Daily Blank, the number pif
responses and views, saved as post meta data. This enables a new short
code [topdaily] to list in a page or widget. And tweaked the use of
page-debug.php as a place holder for one time fix code, see the ReadMe
for links to gists with sample code